requirements.txt
streamlit
torch
torchvision
numpy
pandas
opencv-python
Pillow
gdown